fixed some minors. trim input strings and add some types

+++ func_1.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 int f(void) {
-signed int var1 = 42;
-signed int var2 =42; 
+long var1 = 42;
+long var2 =42; 
 std::cout << var1+ var2 << std::endl; 
 std::cout << var1- var2 << std::endl; 
 std::cout << var1* var2 << std::endl;
